Added Sam3x DMAC support, split SamDM into 3 files
